// frontend/src/types/propertyRequestStatus.ts
/**
 * Generated by orval v7.11.2 🍺
 * Do not edit manually.
 * OpenAPI definition
 * OpenAPI spec version: v0
 */

export type PropertyRequestStatus = typeof PropertyRequestStatus[keyof typeof PropertyRequestStatus];


// eslint-disable-next-line @typescript-eslint/no-redeclare
export const PropertyRequestStatus = {
  AVAILABLE: 'AVAILABLE',
  RENTED: 'RENTED',
  SOLD: 'SOLD',
  MAINTENANCE: 'MAINTENANCE',
  UNDER_CONSTRUCTION: 'UNDER_CONSTRUCTION',
  OFF_MARKET: 'OFF_MARKET',
} as const;